Delete MESSAGE_SCRIPT as it has been superseded by MESSAGE
